feat: added @ManyToOne for field owner in class Item

+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.item.dto.ItemDto;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto createItem(ItemDto itemDto, Integer userId);
+    ItemDto createItem(ItemDto itemDto, Long userId);
 
-    ItemDto updateItem(ItemDto itemDto, Integer itemId, Integer userId);
+    ItemDto updateItem(ItemDto itemDto, Long itemId, Long userId);
 
-    ItemDto getItemById(Integer itemId, Integer userId);
+    ItemDto getItemById(Long itemId, Long userId);
 
-    List<ItemDto> getAllItems(Integer userId);
+    List<ItemDto> getAllItems(Long userId);
 
     List<ItemDto> searchItemsByText(String text);
 }
